requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
langchain
langchain-openai
PyMuPDF
requests
fastapi-cors
